updated and completed Shape class and test
