+++ lib/Circle.js
@@ -1,4 +1,4 @@
-const Shape = require('./Shapes.js');
+const Shape = require('./Shape.js');
 
 class Circle extends Shape {
     constructor(text, textcolor, shape, shapecolor){
