+++ lib/Rectangle.js
@@ -1,4 +1,4 @@
-const Shape = require('./Shapes.js');
+const Shape = require('./Shape.js');
 
 class Rectangle extends Shape {
     constructor(text, textcolor, shape, shapecolor){
