+++ lib/Shape.js
@@ -1,0 +1,11 @@
+class Shape {
+    constructor(text, textcolor, shape, shapecolor){
+        this.text = text;
+        this.textcolor = textcolor;
+        this.shape = shape;
+        this.shapecolor = shapecolor;
+    }
+}
+
+module.exports = Shape;
+
